Hopefully fixed combo points having an incorrect maximum
Sometimes combo points can get stuck in a state where only 5 will draw on the screen even if the player has 6 max available. It's been really tough to reproduce, but I think the problem is that the frames are getting created but never having a texture set to them. This should fix that.