Speculative fix for reported problem
There's a case where custom counters seemingly get stuck with an outdated maximum number of charges (perhaps only when augmented by a talent such as with Throw Glaive). This may catch and fix it (I couldn't reproduce the problem with a starter DH since you need level 108 or something to get the second charge for Throw Glaive).