Fix CC and Invuln not showing immediately
This is the same fix as was applied to RollTheBones in commit 46883e355b746c4f4d369bb82b74d3ef2bf7e5d2 but I failed to audit the rest of the mod for places where this was possible. This one was reported over on the Curse addon page comments.