Add module showing Energy for Druids
This shows the player's current Energy amount when they're in a form that doesn't use Energy. For example, this could show the user how much energy they currently have while they're in Bear form so that they can shift away and wait on their energy to refill.

Note: requires a new version of LibDogTag-Unit for the bar text tags to function if DogTags are enabled.