Fix popup menu replacement code error on 11.0.2 (temp)
This removes the menu replacement code that is preventing errors when trying to focus someone or dismiss a pet through a right-click menu on an addon. There is a new API, UnitPopup_OpenMenu, that is a replacement for ShowMenu, but it's not a drop-in replacement, so until I have access to the beta or an 11.0.2 PTR there's not much I can do here.