Fix spell cooldown bars in 11.x
GetSpellCooldown used to return 1 for "has duration", but the new native version returns true. Fun.

Additionally, it looks like the events triggering UpdateCustomBarEvent() are passing a nil argument that the mod thought was supposed to be the affected unit. So I also fixed that, though I don't know if that's an 11.x thing or not.

https://www.wowace.com/projects/ice-hud/issues/363