Convert bars to be StatusBar frames
This does not work everywhere yet, but it does get the most basic functionality of bars filling as expected in Midnight. More to come.

I would _really_ like to find a way to not have to duplicate every texture for a left/right variant, especially because if I want to support horizontal bars, which StatusBars should make much much simpler, I would need even more variants. There must be a way to make this work programmatically, but I wasn't having any luck attempting to flip or SetTexCoord or anything on these since the StatusBar was forcing the coords back every frame and we'd see it flickering in both places...