Fix target castbar showing wrong durations
Since we can't access spellcast start and end times (secret values), we can't mathematically determine when the spell will end or how long it will take to cast, so we can only show that a cast is ongoing.

Also fixed normal completed spellcasts not fading out as intended if the flashInstants mode was set to Never; we were overzelously treating all Flash calls as instant spells when they weren't necessarily.