Fix cast lag and other non-secret-related bar placements
Now we can choose to create a StatusBar- or plain-ol-Frame-based bar, the latter of which we have much more control over the placement and appearance of but at the cost of no Secret-related functionality support.

This also contains initial support for rotated StatusBars that actually behave correctly. The backgrounds and aforementioned non-StatusBar portions are not doing the right thing right now, but that's a problem for later...maybe. Or, more likely, rotation is disabled entirely going forward.